Improve fault handling.
ReadRTD no longer shifts out the fault flag when reading the temperature
register, so callers can make use of it. Instead the temperature call
extracts the flag and returns it. Note that this changes the ReadRTD
return value.

When starting, also set the fault limits to 0x0000 and 0xffff. They are
supposed to be set to that at power-on but my high threshold was for
some reason getting set to 0, so it was constantly alarming.